+++ R/get_data.R
@@ -1,12 +1,13 @@
 #' Get remote data
 #'
 #' Download remotely stored data via \link[piggyback]{pb_download}.
+#' @param save_dir Directory to save data to.
 #' @keywords internal
 #' @inheritParams piggyback::pb_download
 #' @importFrom tools R_user_dir
 get_data <- function(fname,
                      repo = "neurogenomics/MultiEWCE",
-                     storage_dir = tools::R_user_dir(
+                     save_dir = tools::R_user_dir(
                        package = "MultiEWCE",
                        which = "cache"
                      ),
@@ -17,11 +18,11 @@ get_data <- function(fname,
   requireNamespace("piggyback")
   Sys.setenv("piggyback_cache_duration" = 10)
 
-  tmp <- file.path(storage_dir, fname)
-  dir.create(storage_dir, showWarnings = FALSE, recursive = TRUE)
+  tmp <- file.path(save_dir, fname)
+  dir.create(save_dir, showWarnings = FALSE, recursive = TRUE)
   piggyback::pb_download(
     file = fname,
-    dest = storage_dir,
+    dest = save_dir,
     repo = repo,
     overwrite = overwrite,
     tag = tag